pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "pydantic.*",
    "sqlalchemy.*",
    "alembic.*",
    "uvicorn.*",
    "fastapi.*"
]
disallow_untyped_defs = false

[project]
name = "coboarding"
version = "0.1.0"
description = "AI-Powered Job Application Automation for SME Tech Companies in Europe"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.32.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "aiosqlite>=0.19.0",
    "ollama>=0.1.7",
    "spacy>=3.7.4",
    "transformers>=4.38.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "sentence-transformers>=2.5.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "weasyprint>=60.1",
    "markdown>=3.5.2",
    "playwright>=1.42.0",
    "selenium>=4.18.1",
    "undetected-chromedriver>=3.5.4",
    "opencv-python-headless>=4.9.0.80",
    "easyocr>=1.7.1",
    "requests>=2.31.0",
    "python-telegram-bot>=20.8",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
    "cryptography>=42.0.5",
    "redis>=5.0.0",
    "asyncpg>=0.29.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "black>=24.3.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "flake8-bugbear>=24.1.0",
    "flake8-comprehensions>=3.14.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "watchfiles>=0.21.0",
    "ipython>=8.18.0",
    "ipdb>=0.13.13",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.12",
    "mkdocstrings[python]>=0.24.0",
    "types-requests>=2.31.0.20240125",
    "types-python-dateutil>=2.8.19.14",
    "types-pyyaml>=6.0.12.12"
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["app/tests/*", "app/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"
title = "coBoarding Test Coverage"

[tool.coverage.xml]
output = "coverage.xml"
